chatbot-backend/requirements.txt
flask
flask-cors
openai
faiss-cpu
numpy
